+++ changeLog.md
@@ -1,5 +1,16 @@
 # Change logs
 
+## Sprint 2
+### New
+- เพิ่ม class GetDataReport เพื่อ query ข้อมูลออกไปปริน
+- ใน class GetDataReport เพิ่ม function ส่งออกข้อมูลไปเพื่อไปปริ้น โดยมี fucntion ส่งออกดั้งนี้ getAuthorData, getPaperData, getOtherWorkData,getBookData
+- เพิ่ม document ของวิธีใช้ class GetDataReport
+### Update
+- ปรับ function getPaperOpenAlxe ใน class PublicationRetrieval โดยแยกส่วนหีวออกมาสร้าง functionใหม่่ คือ getDataOpenAlex
+### Bug fix
+- แก้ปัญหา getPaperOpenAlxe หาข้อมูลไม่เจอ
+
+
 ## Sprint 1
 ### New
  - เพิ่ม class PublicationRetrieval เพื่อจัดการดึงข้อมูลงานวิจัย